Add network docs (#35)
Fix #24 

- Add documentation about the network tracking and how to use it in the
benchmarks
- Update the ``NetworkTracker`` class to track the total time directly,
rather than having the ``network_activity_tracker`` track the time
separately. This change addresses the following issues: a)
``NetworkTracker`` was missing the total time so results are not the
same compared to ``network_activity_tracker`` and 2)
``NetworkTracker.asv_network_statistics`` was not being updated in the
``network_activity_tracker`` so the timing result was not being
recorded.
- Update ``NetworkTracker`` and ``network_activity_tracker`` to allow
the user to optionally set the process ID to track. This will be useful
if/when we need to run code we want to profile in a separate process
(e.g., when running in node.js)

---------

Co-authored-by: Cody Baker <51133164+CodyCBakerPhD@users.noreply.github.com>