Allow user to specify network interface to run tshark on (#113)
The default network interface that tshark selected for my machine was
not the internet one, so no data was tracked. This PR allows the user to
specify an environment variable `NWB_BENCHMARKS_NETWORK_INTERFACE` with
the name of the interface, which will be different on each machine. It
does not seem straightforward to set this automatically. I did not set
this up so that users cannot run the benchmarks if they do not have this
set. It will still use the default network interface in that case.