Updates to flask app and database updating (#153)
Some minor documentation and code changes that came up when trying to
debug the database updating locally.

When debugging, I inadvertently pushed the latest results to the
nwb-benchmarks-database GitHub repo - I updated the app script to skip
pushing new files when in debug mode but let me know your thoughts on if
we need any better protections.

From what I could tell, the `repackage_as_parquet` function was not
saving the values from the network tests - which are stored as a
dictionary of several metrics with their associated values. I added an
additional column / attribute to the results that saves the variable
name of each metric (and in the case of the timing tests, uses the
variable name "time")

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Cody Baker <51133164+CodyCBakerPhD@users.noreply.github.com>